refactor: relocate LogDetailsPanel components to pipeline-logs directory - components/log-details-panel/ → app/components/pipeline-logs/LogDetailsPanel/ 이동 - PipelineLogsTable.tsx의 import 경로 업데이트 - 파이프라인 로그 관련 모든 컴포넌트들을 동일한 디렉토리 하위로 통합 - 기존 기능 및 컴포넌트 구조는 변경 없이 유지
