+++ app/components/pipeline-logs/LogDetailsPanel/BuildSummary.tsx
