+++ app/components/pipeline-logs/LogDetailsPanel/LogDetailsPanel.tsx
