+++ app/components/pipeline-logs/LogDetailsPanel/LogExpandedView.tsx
