+++ app/components/pipeline-logs/LogDetailsPanel/LogPreview.tsx
