+++ app/components/pipeline-logs/LogDetailsPanel/LogSummaryView.tsx
