+++ app/components/pipeline-logs/LogDetailsPanel/LogViewer.tsx
