+++ app/components/pipeline-logs/LogDetailsPanel/demo.tsx
