+++ app/components/pipeline-logs/LogDetailsPanel/hooks/useKeyboardShortcuts.ts
