+++ app/components/pipeline-logs/LogDetailsPanel/hooks/useLogSearch.ts
