+++ app/components/pipeline-logs/LogDetailsPanel/index.ts
