+++ app/components/pipeline-logs/LogDetailsPanel/mockData.ts
