+++ app/components/pipeline-logs/LogDetailsPanel/types.ts
