+++ app/components/pipeline-logs/LogDetailsPanel/utils/logFormatters.ts
