+++ app/components/pipeline-logs/LogDetailsPanel/utils/timeUtils.ts
